Charge NanoGPT usage at 0.2p instead of a 1p ledger floor.
Ledger amounts now allow tenths of a penny so NanoGPT floors can bill accurately while other vendors keep the 1p minimum.

+++ app/Exceptions/InsufficientCreditsException.php
@@ -7,8 +7,8 @@ use RuntimeException;
 class InsufficientCreditsException extends RuntimeException
 {
     public function __construct(
-        public readonly int $requiredPence = 0,
-        public readonly int $balancePence = 0,
+        public readonly float $requiredPence = 0,
+        public readonly float $balancePence = 0,
         string $message = 'Your balance must be more than 20p to run this. Subscribe to the platform plan for monthly credit value, or top up on the Billing page.',
     ) {
         parent::__construct($message);
